+++ src/Proto3/Wire/Reverse.hs
@@ -28,6 +28,7 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE MagicHash #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
 module Proto3.Wire.Reverse
     ( -- * `BuildR` type
@@ -105,12 +106,12 @@ import qualified Data.Text.Short               as TS
 import           Data.Vector.Generic           ( Vector )
 import           Data.Word                     ( Word8, Word16, Word32, Word64 )
 import           Foreign                       ( castPtr, copyBytes )
-import           GHC.Exts                      ( Addr#, Int(..), Int# )
+import           GHC.Exts                      ( Addr#, Int(..), Int#, Proxy#, proxy# )
 #if !MIN_VERSION_bytestring(0,11,0)
 import           GHC.Exts                      ( plusAddr# )
 #endif
 import           GHC.ForeignPtr                ( ForeignPtr(..), ForeignPtrContents )
-import           GHC.TypeLits                  ( KnownNat )
+import           GHC.TypeLits                  ( KnownNat, natVal' )
 import           Proto3.Wire.Encode.Repeated   ( Repeated(..), ToRepeated(..) )
 import           Proto3.Wire.Reverse.Internal
 import qualified Proto3.Wire.Reverse.Prim      as Prim
@@ -871,7 +872,7 @@ vectorBuildR f = etaBuildR (foldlRVector (\acc x -> acc <> f x) mempty)
 --
 -- See also: 'repeatedFixedPrimR', 'vectorBuildR'
 repeatedBuildR :: ToRepeated c BuildR => c -> BuildR
-repeatedBuildR = etaBuildR (foldr (flip (<>)) mempty . reverseRepeated . toRepeated)
+repeatedBuildR = etaBuildR (\xs -> foldMapRepeated (toRepeated xs) id)
 {-# INLINE repeatedBuildR #-}
 
 -- | Concatenates the given fixed-width primitives, iterating right to left where practical
@@ -883,12 +884,15 @@ repeatedBuildR = etaBuildR (foldr (flip (<>)) mempty . reverseRepeated . toRepea
 -- [42,67]
 --
 -- See also: 'repeatedBuildR'
-repeatedFixedPrimR :: (ToRepeated c (Prim.FixedPrim w), KnownNat w) => c -> BuildR
+repeatedFixedPrimR :: forall c w . (ToRepeated c (Prim.FixedPrim w), KnownNat w) => c -> BuildR
 repeatedFixedPrimR = etaBuildR $ \c ->
-  let ReverseRepeated prediction prims = toRepeated c in
-  case prediction of
-    Nothing -> foldr (\p a -> a <> Prim.liftBoundedPrim (Prim.liftFixedPrim p)) mempty prims
-    Just count -> Prim.unsafeReverseFoldMapFixedPrim id count prims
+  let MkRepeated count prims = toRepeated c in
+  case count of
+    Left _ ->
+      prims (\p -> Prim.liftBoundedPrim (Prim.liftFixedPrim p))
+    Right n ->
+      let w = fromInteger (natVal' (proxy# :: Proxy# w))
+      in ensure (w * n) (prims (\p -> Prim.unsafeBuildBoundedPrim (Prim.liftFixedPrim p)))
 {-# INLINE repeatedFixedPrimR #-}
 
 -- | Exported for testing purposes only.